fix(filesystem): fix numerous problems with searching

+++ lua/neo-tree/sources/filesystem/lib/filter_external.lua
@@ -2,6 +2,7 @@ local vim = vim
 local log = require("neo-tree.log")
 local Job = require("plenary.job")
 local utils = require("neo-tree.utils")
+local Queue = require("neo-tree.collections").Queue
 
 local M = {}
 local fd_supports_max_results = nil
@@ -43,20 +44,25 @@ local get_find_command = function(state)
   return state.find_command
 end
 
-local running_jobs = {}
-M.cancel = function()
-  while #running_jobs > 0 do
-    local job = table.remove(running_jobs, 1)
-    local pid = job.pid
-    job:shutdown()
-    if pid ~= nil then
-      if utils.is_windows then
-        vim.fn.system("taskkill /F /T /PID " .. pid)
-      else
-        vim.fn.system("kill -9 " .. pid)
-      end
+local running_jobs = Queue:new()
+local kill_job = function(job)
+  local pid = job.pid
+  job:shutdown()
+  if pid ~= nil and pid > 0 then
+    if utils.is_windows then
+      vim.fn.system("taskkill /F /T /PID " .. pid)
+    else
+      vim.fn.system("kill -9 " .. pid)
     end
   end
+  return true
+end
+
+M.cancel = function()
+  if running_jobs:is_empty() then
+    return
+  end
+  running_jobs:for_each(kill_job)
 end
 
 ---@class FileTypes
@@ -84,9 +90,6 @@ end
 ---@param on_insert? fun(err: string, line: string): any Executed for each line of stdout and stderr.
 ---@param on_exit? fun(return_val: table): any Executed at the end.
 M.filter_files_external = function(cmd, path, glob, regex, full_path, types, ignore, limit, find_args, on_insert, on_exit)
-  -- First cancel the last job if it is still running
-  M.cancel()
-
   if glob ~= nil and regex ~= nil then
     local log_msg = string.format([[glob: %s, regex: %s]], glob, regex)
     log.warn("both glob and regex are set. glob will take precedence. " .. log_msg)
@@ -111,6 +114,22 @@ M.filter_files_external = function(cmd, path, glob, regex, full_path, types, ign
     for _, v in pairs({ ... }) do
       if v ~= nil then
         args[#args + 1] = v
+      end
+    end
+  end
+
+  local function append_find_args()
+    if find_args then
+      if type(find_args) == "string" then
+        append(find_args)
+      elseif type(find_args) == "table" then
+        if find_args[1] then
+          append(unpack(find_args))
+        elseif find_args[cmd] then
+          append(unpack(find_args[cmd])) ---@diagnostic disable-line
+        end
+      elseif type(find_args) == "function" then
+        args = find_args(cmd, path, glob, args)
       end
     end
   end
@@ -142,6 +161,7 @@ M.filter_files_external = function(cmd, path, glob, regex, full_path, types, ign
     if glob ~= nil then
       append("--glob")
     end
+    append_find_args()
     append("--", glob or regex or "")
     append(path)
   elseif cmd == "find" then
@@ -173,28 +193,17 @@ M.filter_files_external = function(cmd, path, glob, regex, full_path, types, ign
     elseif regex ~= nil then
       append("-regextype", "sed", "-regex", regex)
     end
+    append_find_args()
   elseif cmd == "fzf" then
     -- This does not work yet, there's some kind of issue with how fzf uses stdout
     error("fzf is not a supported find_command")
+    append_find_args()
     append("--no-sort", "--no-expect", "--filter", glob or regex) -- using the raw term without glob patterns
   elseif cmd == "where" then
+    append_find_args()
     append("/r", path, glob or regex)
   else
     return { "No search command found!" }
-  end
-
-  if find_args then
-    if type(find_args) == "string" then
-      append(find_args)
-    elseif type(find_args) == "table" then
-      if find_args[1] then
-        append(unpack(find_args))
-      elseif find_args[cmd] then
-        append(unpack(find_args[cmd])) ---@diagnostic disable-line
-      end
-    elseif type(find_args) == "function" then
-      args = find_args(cmd, path, glob, args)
-    end
   end
 
   if fd_supports_max_results then
@@ -224,9 +233,11 @@ M.filter_files_external = function(cmd, path, glob, regex, full_path, types, ign
       end
     end,
   })
-  M.cancel()
+
+  -- This ensures that only one job is running at a time
+  running_jobs:for_each(kill_job)
+  running_jobs:add(job)
   job:start()
-  table.insert(running_jobs, job)
 end
 
 local function fzy_sort_get_total_score(terms, path)
@@ -274,12 +285,16 @@ M.fzy_sort_files = function(opts, state)
   -- The fzy score is then used to sort the results
   local chars = {}
   local regex = ".*"
+  local chars_to_escape = { "%", "+", "-", "?", "[", "^", "$", "(", ")", "{", "}", "=", "!", "<", ">", "|", ":", "#" }
   for _, term in ipairs(terms) do
     for c in term:gmatch(".") do
       if not chars[c] then
-        regex = regex .. c .. "+.*"
+        chars[c] = true
+        if chars_to_escape[c] then
+          c = [[\]] .. c
+        end
+        regex = regex .. c .. ".*"
       end
-      chars[c] = true
     end
   end
 
@@ -289,9 +304,6 @@ M.fzy_sort_files = function(opts, state)
   state.fzy_sort_result_scores = {}
   local function on_insert(err, path)
     if not err then
-      if result_counter >= limit then
-        return
-      end
       local relative_path = path
       if not full_path_words and #path > pwd_length and path:sub(1, pwd_length) == pwd then
         relative_path = "./" .. path:sub(pwd_length + 1)
@@ -307,6 +319,9 @@ M.fzy_sort_files = function(opts, state)
         result_counter = result_counter + 1
         modify_parent_scores(state.fzy_sort_result_scores, path, score)
         opts.on_insert(nil, path)
+        if result_counter >= limit then
+          vim.schedule(M.cancel)
+        end
       end
     end
   end
@@ -315,7 +330,7 @@ M.fzy_sort_files = function(opts, state)
     { directory = fuzzy_finder_mode == "directory", file = fuzzy_finder_mode ~= "directory" },
     { dotfiles = not filters.visible and filters.hide_dotfiles,
       gitignore = not filters.visible and filters.hide_gitignored },
-    limit * 10, opts.find_args, on_insert, opts.on_exit)
+    nil, opts.find_args, on_insert, opts.on_exit)
 end
 
 M.find_files = function(opts)
